Metodo Clientes implementado

+++ src/main/java/br/com/edward/restfull/Service/impl/PessoaServiceImpl.java
@@ -1,21 +1,23 @@
 package br.com.edward.restfull.service.impl;
 
 import java.util.List;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import br.com.edward.restfull.domain.Pessoa;
 import br.com.edward.restfull.repository.PessoaRepository;
 import br.com.edward.restfull.service.PessoaService;
 
+@Transactional
 @Service
 public class PessoaServiceImpl implements PessoaService {
 
-	@Autowired
+    @Autowired
     private PessoaRepository pessoaRepository;
     
-
     @Override
     public Pessoa consultar(Long idPessoa) {
         return pessoaRepository.findById(idPessoa).orElse(null);
@@ -29,7 +31,9 @@ public class PessoaServiceImpl implements PessoaService {
     @Override
     public Pessoa remover(Long id) {
         Pessoa pessoaRemover = this.consultar(id);
-        pessoaRepository.delete(pessoaRemover);
+        if (Objects.nonNull(pessoaRemover)) {
+            pessoaRepository.delete(pessoaRemover);
+        }
         return pessoaRemover;
     }
 }
